refactor: move extension related code to io.substrait.extension
