+++ core/src/main/java/io/substrait/expression/AbstractFunctionInvocation.java
@@ -1,6 +1,6 @@
 package io.substrait.expression;
 
-import io.substrait.function.SimpleExtension;
+import io.substrait.extension.SimpleExtension;
 import io.substrait.type.Type;
 import java.util.List;
