+++ core/src/main/java/io/substrait/expression/EnumArg.java
@@ -1,6 +1,6 @@
 package io.substrait.expression;
 
-import io.substrait.function.SimpleExtension;
+import io.substrait.extension.SimpleExtension;
 import java.util.Optional;
 import org.immutables.value.Value;
