+++ core/src/main/java/io/substrait/extension/ExtensionLookup.java
@@ -1,6 +1,4 @@
-package io.substrait.expression;
-
-import io.substrait.function.SimpleExtension;
+package io.substrait.extension;
 
 /**
  * Interface with operations for resolving references to {@link
@@ -8,7 +6,6 @@ import io.substrait.function.SimpleExtension;
  * functions or types.
  */
 public interface ExtensionLookup {
-  // TODO: Move to io.substrait.extension
   SimpleExtension.ScalarFunctionVariant getScalarFunction(
       int reference, SimpleExtension.ExtensionCollection extensions);
