+++ core/src/main/java/io/substrait/type/proto/ParameterizedProtoConverter.java
@@ -1,6 +1,6 @@
 package io.substrait.type.proto;
 
-import io.substrait.expression.proto.ExtensionCollector;
+import io.substrait.extension.ExtensionCollector;
 import io.substrait.function.TypeExpression;
 import io.substrait.function.TypeExpressionVisitor;
 import io.substrait.proto.ParameterizedType;
