+++ core/src/test/java/io/substrait/relation/ProtoRelConverterTest.java
@@ -5,8 +5,8 @@ import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import io.substrait.dsl.SubstraitBuilder;
-import io.substrait.expression.proto.ExtensionCollector;
-import io.substrait.function.SimpleExtension;
+import io.substrait.extension.ExtensionCollector;
+import io.substrait.extension.SimpleExtension;
 import io.substrait.io.substrait.extension.AdvancedExtension;
 import io.substrait.relation.utils.StringHolder;
 import java.io.IOException;
