+++ core/src/test/java/io/substrait/relation/utils/StringHolderHandlingProtoRelConverter.java
@@ -3,8 +3,8 @@ package io.substrait.relation.utils;
 import com.google.protobuf.Any;
 import com.google.protobuf.InvalidProtocolBufferException;
 import com.google.protobuf.StringValue;
-import io.substrait.expression.ExtensionLookup;
-import io.substrait.function.SimpleExtension;
+import io.substrait.extension.ExtensionLookup;
+import io.substrait.extension.SimpleExtension;
 import io.substrait.relation.Extension;
 import io.substrait.relation.ProtoRelConverter;
