+++ core/src/test/java/io/substrait/type/proto/AggregateRoundtripTest.java
@@ -6,8 +6,8 @@ import io.substrait.expression.AggregateFunctionInvocation;
 import io.substrait.expression.Expression;
 import io.substrait.expression.ExpressionCreator;
 import io.substrait.expression.ImmutableExpression;
-import io.substrait.expression.proto.ExtensionCollector;
-import io.substrait.function.SimpleExtension;
+import io.substrait.extension.ExtensionCollector;
+import io.substrait.extension.SimpleExtension;
 import io.substrait.proto.AggregateFunction;
 import io.substrait.relation.Aggregate;
 import io.substrait.relation.ImmutableAggregate;
