+++ core/src/test/java/io/substrait/type/proto/ExtensionRoundtripTest.java
@@ -4,8 +4,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import io.substrait.dsl.SubstraitBuilder;
 import io.substrait.expression.Expression;
-import io.substrait.expression.proto.ExtensionCollector;
-import io.substrait.function.SimpleExtension;
+import io.substrait.extension.ExtensionCollector;
+import io.substrait.extension.SimpleExtension;
 import io.substrait.io.substrait.extension.AdvancedExtension;
 import io.substrait.relation.Aggregate;
 import io.substrait.relation.Cross;
