+++ core/src/test/java/io/substrait/type/proto/LocalFilesRoundtripTest.java
@@ -6,8 +6,8 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import io.substrait.expression.ExpressionCreator;
 import io.substrait.expression.FieldReference;
 import io.substrait.expression.ImmutableFieldReference;
-import io.substrait.expression.proto.ExtensionCollector;
-import io.substrait.function.SimpleExtension;
+import io.substrait.extension.ExtensionCollector;
+import io.substrait.extension.SimpleExtension;
 import io.substrait.proto.ReadRel;
 import io.substrait.relation.LocalFiles;
 import io.substrait.relation.ProtoRelConverter;
