+++ core/src/test/java/io/substrait/type/proto/TestTypeRoundtrip.java
@@ -2,8 +2,8 @@ package io.substrait.type.proto;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import io.substrait.expression.proto.ExtensionCollector;
-import io.substrait.function.ImmutableSimpleExtension;
+import io.substrait.extension.ExtensionCollector;
+import io.substrait.extension.ImmutableSimpleExtension;
 import io.substrait.type.Type;
 import io.substrait.type.TypeCreator;
 import org.junit.jupiter.params.ParameterizedTest;
