+++ isthmus/src/main/java/io/substrait/isthmus/expression/AggregateFunctionConverter.java
@@ -5,7 +5,7 @@ import io.substrait.expression.AggregateFunctionInvocation;
 import io.substrait.expression.Expression;
 import io.substrait.expression.ExpressionCreator;
 import io.substrait.expression.FunctionArg;
-import io.substrait.function.SimpleExtension;
+import io.substrait.extension.SimpleExtension;
 import io.substrait.isthmus.SubstraitRelVisitor;
 import io.substrait.proto.AggregateFunction;
 import io.substrait.type.Type;
