+++ isthmus/src/main/java/io/substrait/isthmus/expression/FunctionConverter.java
@@ -4,8 +4,8 @@ import com.google.common.collect.*;
 import io.substrait.expression.Expression;
 import io.substrait.expression.ExpressionCreator;
 import io.substrait.expression.FunctionArg;
+import io.substrait.extension.SimpleExtension;
 import io.substrait.function.ParameterizedType;
-import io.substrait.function.SimpleExtension;
 import io.substrait.function.ToTypeString;
 import io.substrait.isthmus.TypeConverter;
 import io.substrait.isthmus.Utils;
